[Gordon]fix user data statistics

+++ src/main/java/com/ita/domain/dto/suadmin/UserData.java
@@ -11,5 +11,5 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class UserData {
-  private List<Long> data;
+  private List<Integer> data;
 }
